+++ tests/settings.py
@@ -171,6 +171,7 @@ INSTALLED_APPS = (
     "allauth.socialaccount.providers.strava",
     "allauth.socialaccount.providers.stripe",
     "allauth.socialaccount.providers.telegram",
+    "allauth.socialaccount.providers.tiktok",
     "allauth.socialaccount.providers.trainingpeaks",
     "allauth.socialaccount.providers.trello",
     "allauth.socialaccount.providers.tumblr",